requirements.txt
# Core
boto3>=1.34
duckdb>=0.10
pyarrow>=15.0
pandas>=2.1
pyyaml>=6.0
click>=8.1
python-dotenv>=1.0

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
moto[s3]>=5.0
